# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Head Coach Dream Save API
  version: 1.0.0
servers:
  - url: http://localhost:8787
paths:
  /api/v1/health:
    get:
      summary: Health check
      operationId: getHealth
      responses:
        '200':
          description: Healthy
          headers:
            X-Api-Version:
              $ref: '#/components/headers/ApiVersion'
            X-Contract-Version:
              $ref: '#/components/headers/ContractVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                default:
                  value:
                    status: ok
  /api/v1/saves/metadata:
    get:
      summary: List save metadata
      operationId: listSaveMetadata
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Save metadata list
          headers:
            X-Api-Version:
              $ref: '#/components/headers/ApiVersion'
            X-Contract-Version:
              $ref: '#/components/headers/ContractVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataListResponse'
              examples:
                default:
                  value:
                    saves:
                      - saveId: career-demo
                        coachName: Coach Carter
                        teamName: ATLANTA_APEX
                        season: 1
                        week: 1
                        record:
                          wins: 0
                          losses: 0
                        updatedAt: 1710000000000
                        lastPlayed: 1710000000000
                        careerStage: PRE_SEASON
                        version: 1
                    pagination:
                      limit: 50
                      nextCursor: null
                      total: 1
  /api/v1/saves/snapshots:
    post:
      summary: Create a save snapshot
      operationId: createSaveSnapshot
      parameters:
        - $ref: '#/components/parameters/OperationIdHeader'
        - $ref: '#/components/parameters/SequenceNumberHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSnapshotRequest'
      responses:
        '201':
          description: Snapshot created
          headers:
            X-Api-Version:
              $ref: '#/components/headers/ApiVersion'
            X-Contract-Version:
              $ref: '#/components/headers/ContractVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/saves/{saveId}/snapshot:
    get:
      summary: Get a save snapshot
      operationId: getSaveSnapshot
      parameters:
        - $ref: '#/components/parameters/SaveId'
      responses:
        '200':
          description: Snapshot found
          headers:
            X-Api-Version:
              $ref: '#/components/headers/ApiVersion'
            X-Contract-Version:
              $ref: '#/components/headers/ContractVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveSnapshotResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Upsert a save snapshot
      operationId: upsertSaveSnapshot
      parameters:
        - $ref: '#/components/parameters/SaveId'
        - $ref: '#/components/parameters/OperationIdHeader'
        - $ref: '#/components/parameters/SequenceNumberHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertSnapshotRequest'
      responses:
        '200':
          description: Snapshot saved
          headers:
            X-Api-Version:
              $ref: '#/components/headers/ApiVersion'
            X-Contract-Version:
              $ref: '#/components/headers/ContractVersion'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MutationResponse'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      summary: Delete a save snapshot
      operationId: deleteSaveSnapshot
      parameters:
        - $ref: '#/components/parameters/SaveId'
        - $ref: '#/components/parameters/OperationIdHeader'
        - $ref: '#/components/parameters/SequenceNumberHeader'
      responses:
        '204':
          description: Snapshot deleted
          headers:
            X-Api-Version:
              $ref: '#/components/headers/ApiVersion'
            X-Contract-Version:
              $ref: '#/components/headers/ContractVersion'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  headers:
    ApiVersion:
      description: Backward-compatible API major version.
      schema:
        type: string
        const: v1
    ContractVersion:
      description: Contract release version consumed by mobile clients.
      schema:
        type: string
        pattern: '^\\d{4}-\\d{2}-\\d{2}$'
  parameters:
    SaveId:
      name: saveId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
    OperationIdHeader:
      name: x-operation-id
      in: header
      required: true
      schema:
        type: string
        minLength: 1
      description: Unique client-generated operation ID used for idempotent retries.
    SequenceNumberHeader:
      name: x-sequence-number
      in: header
      required: true
      schema:
        type: integer
        minimum: 1
      description: Monotonic sequence per save stream; must be previous + 1.
  responses:
    Conflict:
      description: Operation replay or sequence conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictError'
    NotFound:
      description: Resource was not found
      headers:
        X-Api-Version:
          $ref: '#/components/headers/ApiVersion'
        X-Contract-Version:
          $ref: '#/components/headers/ContractVersion'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            default:
              value:
                error: save not found
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok]
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SaveMetadata:
      type: object
      required:
        [saveId, coachName, teamName, season, week, record, updatedAt, lastPlayed, careerStage, version]
      properties:
        saveId:
          type: string
        coachName:
          type: string
        teamName:
          type: string
        season:
          type: number
        week:
          type: number
        record:
          type: object
          required: [wins, losses]
          properties:
            wins:
              type: number
            losses:
              type: number
        updatedAt:
          type: integer
          description: Unix timestamp in milliseconds.
          minimum: 0
        lastPlayed:
          type: integer
          description: Unix timestamp in milliseconds.
          minimum: 0
        careerStage:
          type: string
        version:
          type: number
    Pagination:
      type: object
      required: [limit, nextCursor, total]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        nextCursor:
          oneOf:
            - type: string
            - type: 'null'
        total:
          type: integer
          minimum: 0
    MetadataListResponse:
      type: object
      required: [saves, pagination]
      properties:
        saves:
          type: array
          items:
            $ref: '#/components/schemas/SaveMetadata'
        pagination:
          $ref: '#/components/schemas/Pagination'
    SaveState:
      type: object
      additionalProperties: true
      description: Existing game save/state object shape from the web app save schema.
    SaveSnapshotResponse:
      type: object
      required: [saveId, metadata, state]
      properties:
        saveId:
          type: string
        metadata:
          $ref: '#/components/schemas/SaveMetadata'
        state:
          $ref: '#/components/schemas/SaveState'
    MutationResponse:
      allOf:
        - $ref: '#/components/schemas/SaveSnapshotResponse'
        - type: object
          required: [operationId, sequence]
          properties:
            operationId:
              type: string
            sequence:
              type: integer
    ConflictError:
      type: object
      required: [error]
      properties:
        error:
          type: string
        expectedSequence:
          type: integer
        actualSequence:
          type: integer
    CreateSnapshotRequest:
      type: object
      required: [state]
      properties:
        saveId:
          type: string
        state:
          $ref: '#/components/schemas/SaveState'
    UpsertSnapshotRequest:
      type: object
      required: [state]
      properties:
        state:
          $ref: '#/components/schemas/SaveState'